Cost and Benefit System
Finally, after 3 months, managed to get around to fixing and adding the Cost and Benefit system. Now all that's left is to balance the cost and benefit of the cards and add winning conditions as well as some sort of tutorial text and explain how to restart the game without quitting and how to exit. Perhaps some music as well and more detail to cards, background, deck, and table.
Important note: The UI freaks out when not playing 16:9 aspect ratio. Not sure why, need to find out and fix.